Add files via upload
Here is notebook that I showcased today for EDA and preprocessing, it excludes the modeling and waveform plotting that I briefly showed in the meeting. You will also need to upload the 2 png images that I provided when you run the notebook. Additionally you will need to change the path of your csv files in cell 2 (I stored mine in a folder named "Station_Data"). In the notebook I detail an issue about the function "Earthquake_Finder", if you come up with a solution it would be great if you could push a commit addressing it.