Add files via upload
Because we previously had issues with downloading models from seisbench I wrote this short script that downloads all of the pretrained models and pickles them into a file on your path. I also included an example of how to load the pickle from the file and work with the nested dictionaries (how I chose to store the models) that it loads.