Modeling Notebook
Provides the framework necessary to run all the models that we are concerned with using on the given streams. Import note: you must have the package 'ipynb' installed in order to properly use the notebook. I use the multiple functions that were created in other notebooks. Yuancong explained in our Slack channel how to install packages if you need that info.